feat: finalizando gateway http de registro
