+++ src/infra/modules/login/login-gateway.ts
@@ -2,4 +2,5 @@ import { LoginRequestDTO } from "./dto/login-request-dto";
 
 export interface LoginGateway {
   login(user: LoginRequestDTO): Promise<void>
+  logout(): Promise<void>
 }
