+++ src/infra/modules/user/dto/get-user-dto.ts
@@ -1,0 +1,7 @@
+import { EnumProfile } from "@/presentation/shared/layout/components/profile/profile";
+
+export interface GetUserDTO {
+  id: string;
+  name: string;
+  role: EnumProfile
+}
