+++ src/infra/modules/user/user-gateway.ts
@@ -1,0 +1,5 @@
+import { GetUserDTO } from "./dto/get-user-dto";
+
+export interface UserGateway {
+  get(): Promise<GetUserDTO>
+}
