+++ src/presentation/shared/layout/components/sidebar.tsx
@@ -1,3 +1,5 @@
+'use client'
+import { loginGatewayHttp } from "@/infra/modules/login/login-gateway-http"
 import {
   Sidebar,
   SidebarContent,
@@ -9,18 +11,35 @@ import {
   SidebarMenuButton,
   SidebarMenuItem,
 } from "@/presentation/external/components/ui/sidebar"
+import { APP_ROUTES } from "@/shared/constants/route"
 
 import { Home, FileText, UsersIcon, BarChart3, LogOut } from "lucide-react"
 import Image from "next/image"
+import { useRouter } from "next/navigation"
+interface HomeSideBarProps {
+  userIsMentor: boolean
+}
 
-const menuItems = [
-  { title: "Home", icon: Home, url: "#", isActive: true },
-  { title: "Editais", icon: FileText, url: "#" },
-  { title: "Mentores", icon: UsersIcon, url: "#" },
-  { title: "Meu progresso", icon: BarChart3, url: "#" },
-]
+export function HomeSideBar({ userIsMentor }: HomeSideBarProps) {
 
-export function HomeSideBar() {
+  const { push } = useRouter()
+
+  const allMenuItems = [
+    { title: "Home", icon: Home, url: "#", isActive: true },
+    { title: "Editais", icon: FileText, url: APP_ROUTES.create_edict, onlyMentor: true },
+    { title: "Mentores", icon: UsersIcon, url: "#", },
+    { title: "Meu progresso", icon: BarChart3, url: "#" },
+  ]
+
+  const menuItems = allMenuItems.filter(item => {
+    if (item.onlyMentor) return userIsMentor
+    return true
+  })
+
+  async function handleLogOut() {
+    await loginGatewayHttp.logout().then(() => push(APP_ROUTES.login))
+  }
+
   return (
     <Sidebar className="border-r-0">
       <SidebarHeader className="p-6">
@@ -53,11 +72,9 @@ export function HomeSideBar() {
       <SidebarFooter className="p-4">
         <SidebarMenu>
           <SidebarMenuItem>
-            <SidebarMenuButton asChild className="hover:bg-red-50 hover:text-red-600">
-              <a href="#" className="flex items-center gap-3">
-                <LogOut className="w-5 h-5" />
-                <span>Sair</span>
-              </a>
+            <SidebarMenuButton className="hover:bg-red-50 hover:text-red-600" onClick={handleLogOut}>
+              <LogOut className="w-5 h-5" />
+              <span>Sair</span>
             </SidebarMenuButton>
           </SidebarMenuItem>
         </SidebarMenu>
