+++ src/shared/constants/route.ts
@@ -1,7 +1,12 @@
 export const APP_ROUTES = {
   login: "/login",
   register: "/registro",
-  home: "/inicio",
+  home: "/",
   create_student_profile: "/profile-student",
   create_mentor_profile: "/profile-mentor",
+  create_edict: "/criar-edital"
 }
+
+export const API_ROUTES = {
+
+}
